i18n for password & line amount normalizer
