+++ app/ui/src/account/change-password-modal.html
@@ -9,7 +9,7 @@
         </header>
         <div class="modal-body">
             <form class="auth-form">
-                <choose-password component.ref="choosePassword" password.two-way="password" confirm-password.two-way="confirmPassword" show-heading.bind="true"></choose-password>
+                <choose-password component.ref="choosePassword" password.two-way="password" confirm-password.two-way="confirmPassword"></choose-password>
                 <p if.bind="error" class="form-error" t="forgot.reset-failed">Change password failed</p>
             </form>
         </div>
