+++ app/ui/src/components/choose-password/choose-password.html
@@ -14,13 +14,13 @@
     </div>
     <template if.bind="password">
         <ul class="pw-hints">
-            <li class.bind="rules.lengthOk ? 'ok' : ''">At least 12 characters</li>
-            <li class.bind="rules.lowerOk ? 'ok' : ''">At least 1 lower</li>
-            <li class.bind="rules.upperOk ? 'ok' : ''">At least 1 UPPER</li>
-            <li class.bind="rules.numberOk ? 'ok' : ''">At least 1 number</li>
-            <li class.bind="rules.symbolOk ? 'ok' : ''">At least 1 symbol</li>
-            <li class.bind="rules.matchOk ? 'ok' : ''">Passwords match</li>
+            <li class.bind="rules.lengthOk ? 'ok' : ''" t="password-rules.length">At least 12 characters</li>
+            <li class.bind="rules.lowerOk ? 'ok' : ''" t="password-rules.lower">At least 1 lowercase letter</li>
+            <li class.bind="rules.upperOk ? 'ok' : ''" t="password-rules.upper">At least 1 uppercase letter</li>
+            <li class.bind="rules.numberOk ? 'ok' : ''" t="password-rules.number">At least 1 number</li>
+            <li class.bind="rules.symbolOk ? 'ok' : ''" t="password-rules.symbol">At least 1 symbol</li>
+            <li class.bind="rules.matchOk ? 'ok' : ''" t="password-rules.match">Passwords match</li>
         </ul>
-        <meter min="0" max="8" low="4" high="7" optimum="8" value.bind="rules.pwScore" aria-label="Password strength"></meter>
+        <meter min="0" max="8" low="4" high="7" optimum="8" value.bind="rules.pwScore" t="[aria-label]password-rules.strength" aria-label="Password strength"></meter>
     </template>
 </template>
